refactor(automation): decompose PMA automation store into domain modules
- Extract pma_automation_domain_translation.py for event-to-rule bridging
- Extract pma_automation_mirror.py for legacy compatibility mirroring
- Extract pma_automation_persistence.py for SQLite CRUD operations
- Extract pma_automation_records.py for normalized record types
- Slim pma_automation_store.py to thin compatibility adapter
