+++ src/codex_autorunner/core/pma_automation_store.py
@@ -1,19 +1,11 @@
 from __future__ import annotations
 
-import json
 import logging
-import sqlite3
-import uuid
-from dataclasses import asdict, dataclass, field
+from dataclasses import asdict
 from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, Optional, cast
+from typing import Any, Optional
 
-from .automation import (
-    AutomationStore,
-    mirror_pma_subscription_rule,
-    mirror_pma_timer_schedule,
-)
 from .chat_bindings import (
     active_chat_binding_metadata_by_thread,
     preferred_non_pma_chat_notification_source_for_workspace,
@@ -23,12 +15,20 @@ from .config_contract import ConfigError
 from .locks import file_lock
 from .managed_thread_store import ManagedThreadStore
 from .orchestration.sqlite import open_orchestration_sqlite
+from .pma_automation_domain_translation import PmaAutomationDomainTranslator
+from .pma_automation_mirror import PmaAutomationMirror
+from .pma_automation_persistence import PmaAutomationPersistence
+from .pma_automation_records import (
+    PmaAutomationTimer,
+    PmaAutomationWakeup,
+    PmaLifecycleSubscription,
+    _resolve_subscription_max_matches,
+)
 from .pma_automation_types import (
     DEFAULT_PMA_LANE_ID,
     DEFAULT_WATCHDOG_IDLE_SECONDS,
     PMA_AUTOMATION_STORE_FILENAME,
     PMA_AUTOMATION_VERSION,
-    TIMER_TYPE_ONE_SHOT,
     TIMER_TYPE_WATCHDOG,
     _iso_after_seconds,
     _iso_now,
@@ -38,7 +38,6 @@ from .pma_automation_types import (
     _normalize_non_negative_int,
     _normalize_positive_int,
     _normalize_text,
-    _normalize_text_list,
     _normalize_timer_type,
     _parse_iso,
     default_pma_automation_state,
@@ -53,7 +52,6 @@ from .pma_domain.automation_reducer import (
     reduce_wakeup_dispatch,
     reduce_wakeup_queued,
 )
-from .pma_domain.models import PmaSubscription, PmaTimer, PmaWakeup
 from .pma_domain.subscription_reducer import (
     ReduceTransitionResult,
     TimerFiredEvent,
@@ -62,7 +60,7 @@ from .pma_domain.subscription_reducer import (
     reduce_transition,
 )
 from .pma_origin import extract_pma_origin_metadata, merge_pma_origin_metadata
-from .text_utils import _normalize_pma_delivery_target, lock_path_for
+from .text_utils import _normalize_pma_delivery_target
 
 logger = logging.getLogger(__name__)
 
@@ -116,399 +114,25 @@ def _repo_scoped_subscription_warning(
     )
 
 
-def _resolve_subscription_max_matches(
-    *,
-    max_matches: Any,
-    notify_once: Any = None,
-    metadata: Optional[dict[str, Any]] = None,
-) -> Optional[int]:
-    resolved_max = _normalize_positive_int(max_matches, fallback=None)
-    if resolved_max is not None:
-        return resolved_max
-    if _normalize_bool(notify_once, fallback=False):
-        return 1
-    if isinstance(metadata, dict) and _normalize_bool(
-        metadata.get("notify_once"), fallback=False
-    ):
-        return 1
-    return None
-
-
-def _canonicalize_subscription_entry(data: dict[str, Any]) -> dict[str, Any]:
-    canonical = dict(data)
-    canonical.pop("notify_once", None)
-    metadata_raw = data.get("metadata")
-    metadata = dict(metadata_raw) if isinstance(metadata_raw, dict) else {}
-    resolved_max = _resolve_subscription_max_matches(
-        max_matches=data.get("max_matches"),
-        notify_once=data.get("notify_once"),
-        metadata=metadata,
-    )
-    metadata.pop("notify_once", None)
-    return {
-        **canonical,
-        "metadata": metadata,
-        "max_matches": resolved_max,
-    }
-
-
-@dataclass
-class PmaLifecycleSubscription:
-    subscription_id: str
-    created_at: str
-    updated_at: str
-    state: str = "active"
-    event_types: list[str] = field(default_factory=list)
-    repo_id: Optional[str] = None
-    run_id: Optional[str] = None
-    thread_id: Optional[str] = None
-    lane_id: str = DEFAULT_PMA_LANE_ID
-    from_state: Optional[str] = None
-    to_state: Optional[str] = None
-    reason: Optional[str] = None
-    idempotency_key: Optional[str] = None
-    max_matches: Optional[int] = None
-    match_count: int = 0
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def create(
-        cls,
-        *,
-        event_types: Optional[list[str]] = None,
-        repo_id: Optional[str] = None,
-        run_id: Optional[str] = None,
-        thread_id: Optional[str] = None,
-        lane_id: Optional[str] = None,
-        from_state: Optional[str] = None,
-        to_state: Optional[str] = None,
-        reason: Optional[str] = None,
-        idempotency_key: Optional[str] = None,
-        max_matches: Optional[int] = None,
-        metadata: Optional[dict[str, Any]] = None,
-    ) -> "PmaLifecycleSubscription":
-        stamp = _iso_now()
-        return cls(
-            subscription_id=str(uuid.uuid4()),
-            created_at=stamp,
-            updated_at=stamp,
-            state="active",
-            event_types=_normalize_text_list(event_types or []),
-            repo_id=_normalize_text(repo_id),
-            run_id=_normalize_text(run_id),
-            thread_id=_normalize_text(thread_id),
-            lane_id=_normalize_lane_id(lane_id),
-            from_state=_normalize_text(from_state),
-            to_state=_normalize_text(to_state),
-            reason=_normalize_text(reason),
-            idempotency_key=_normalize_text(idempotency_key),
-            max_matches=_normalize_positive_int(max_matches, fallback=None),
-            match_count=0,
-            metadata=dict(metadata or {}),
-        )
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "PmaLifecycleSubscription":
-        canonical = _canonicalize_subscription_entry(data)
-        subscription_id = _normalize_text(data.get("subscription_id")) or str(
-            uuid.uuid4()
-        )
-        created_at = _normalize_text(data.get("created_at")) or _iso_now()
-        updated_at = _normalize_text(data.get("updated_at")) or created_at
-        state = _normalize_text(data.get("state")) or "active"
-        max_matches = _normalize_positive_int(
-            canonical.get("max_matches"), fallback=None
-        )
-        match_count = _normalize_non_negative_int(
-            canonical.get("match_count"), fallback=0
-        )
-        if match_count is None:
-            match_count = 0
-        metadata_raw = canonical.get("metadata")
-        metadata: dict[str, Any] = (
-            dict(metadata_raw) if isinstance(metadata_raw, dict) else {}
-        )
-        return cls(
-            subscription_id=subscription_id,
-            created_at=created_at,
-            updated_at=updated_at,
-            state=state.lower(),
-            event_types=_normalize_text_list(canonical.get("event_types") or []),
-            repo_id=_normalize_text(canonical.get("repo_id")),
-            run_id=_normalize_text(canonical.get("run_id")),
-            thread_id=_normalize_text(canonical.get("thread_id")),
-            lane_id=_normalize_lane_id(canonical.get("lane_id")),
-            from_state=_normalize_text(canonical.get("from_state")),
-            to_state=_normalize_text(canonical.get("to_state")),
-            reason=_normalize_text(canonical.get("reason")),
-            idempotency_key=_normalize_text(canonical.get("idempotency_key")),
-            max_matches=max_matches,
-            match_count=int(match_count),
-            metadata=metadata,
-        )
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
-
-
-@dataclass
-class PmaAutomationTimer:
-    timer_id: str
-    due_at: str
-    created_at: str
-    updated_at: str
-    state: str = "pending"
-    fired_at: Optional[str] = None
-    timer_type: str = TIMER_TYPE_ONE_SHOT
-    idle_seconds: Optional[int] = None
-    subscription_id: Optional[str] = None
-    repo_id: Optional[str] = None
-    run_id: Optional[str] = None
-    thread_id: Optional[str] = None
-    lane_id: str = DEFAULT_PMA_LANE_ID
-    from_state: Optional[str] = None
-    to_state: Optional[str] = None
-    reason: Optional[str] = None
-    idempotency_key: Optional[str] = None
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def create(
-        cls,
-        *,
-        due_at: str,
-        timer_type: Optional[str] = None,
-        idle_seconds: Optional[int] = None,
-        subscription_id: Optional[str] = None,
-        repo_id: Optional[str] = None,
-        run_id: Optional[str] = None,
-        thread_id: Optional[str] = None,
-        lane_id: Optional[str] = None,
-        from_state: Optional[str] = None,
-        to_state: Optional[str] = None,
-        reason: Optional[str] = None,
-        idempotency_key: Optional[str] = None,
-        metadata: Optional[dict[str, Any]] = None,
-    ) -> "PmaAutomationTimer":
-        stamp = _iso_now()
-        return cls(
-            timer_id=str(uuid.uuid4()),
-            due_at=due_at,
-            created_at=stamp,
-            updated_at=stamp,
-            state="pending",
-            fired_at=None,
-            timer_type=_normalize_timer_type(timer_type),
-            idle_seconds=_normalize_non_negative_int(idle_seconds),
-            subscription_id=_normalize_text(subscription_id),
-            repo_id=_normalize_text(repo_id),
-            run_id=_normalize_text(run_id),
-            thread_id=_normalize_text(thread_id),
-            lane_id=_normalize_lane_id(lane_id),
-            from_state=_normalize_text(from_state),
-            to_state=_normalize_text(to_state),
-            reason=_normalize_text(reason),
-            idempotency_key=_normalize_text(idempotency_key),
-            metadata=dict(metadata or {}),
-        )
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "PmaAutomationTimer":
-        timer_id = _normalize_text(data.get("timer_id")) or str(uuid.uuid4())
-        created_at = _normalize_text(data.get("created_at")) or _iso_now()
-        updated_at = _normalize_text(data.get("updated_at")) or created_at
-        due_at = _normalize_text(data.get("due_at")) or created_at
-        state = _normalize_text(data.get("state")) or "pending"
-        metadata_raw = data.get("metadata")
-        metadata: dict[str, Any] = (
-            dict(metadata_raw) if isinstance(metadata_raw, dict) else {}
-        )
-        return cls(
-            timer_id=timer_id,
-            due_at=due_at,
-            created_at=created_at,
-            updated_at=updated_at,
-            state=state.lower(),
-            fired_at=_normalize_text(data.get("fired_at")),
-            timer_type=_normalize_timer_type(data.get("timer_type")),
-            idle_seconds=_normalize_non_negative_int(data.get("idle_seconds")),
-            subscription_id=_normalize_text(data.get("subscription_id")),
-            repo_id=_normalize_text(data.get("repo_id")),
-            run_id=_normalize_text(data.get("run_id")),
-            thread_id=_normalize_text(data.get("thread_id")),
-            lane_id=_normalize_lane_id(data.get("lane_id")),
-            from_state=_normalize_text(data.get("from_state")),
-            to_state=_normalize_text(data.get("to_state")),
-            reason=_normalize_text(data.get("reason")),
-            idempotency_key=_normalize_text(data.get("idempotency_key")),
-            metadata=metadata,
-        )
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
-
-
-@dataclass
-class PmaAutomationWakeup:
-    wakeup_id: str
-    created_at: str
-    updated_at: str
-    state: str = "pending"
-    dispatched_at: Optional[str] = None
-    source: str = "automation"
-    repo_id: Optional[str] = None
-    run_id: Optional[str] = None
-    thread_id: Optional[str] = None
-    lane_id: str = DEFAULT_PMA_LANE_ID
-    from_state: Optional[str] = None
-    to_state: Optional[str] = None
-    reason: Optional[str] = None
-    timestamp: Optional[str] = None
-    idempotency_key: Optional[str] = None
-    subscription_id: Optional[str] = None
-    timer_id: Optional[str] = None
-    event_id: Optional[str] = None
-    event_type: Optional[str] = None
-    event_data: dict[str, Any] = field(default_factory=dict)
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def create(
-        cls,
-        *,
-        source: str,
-        repo_id: Optional[str] = None,
-        run_id: Optional[str] = None,
-        thread_id: Optional[str] = None,
-        lane_id: Optional[str] = None,
-        from_state: Optional[str] = None,
-        to_state: Optional[str] = None,
-        reason: Optional[str] = None,
-        timestamp: Optional[str] = None,
-        idempotency_key: Optional[str] = None,
-        subscription_id: Optional[str] = None,
-        timer_id: Optional[str] = None,
-        event_id: Optional[str] = None,
-        event_type: Optional[str] = None,
-        event_data: Optional[dict[str, Any]] = None,
-        metadata: Optional[dict[str, Any]] = None,
-    ) -> "PmaAutomationWakeup":
-        stamp = _iso_now()
-        return cls(
-            wakeup_id=str(uuid.uuid4()),
-            created_at=stamp,
-            updated_at=stamp,
-            state="pending",
-            dispatched_at=None,
-            source=_normalize_text(source) or "automation",
-            repo_id=_normalize_text(repo_id),
-            run_id=_normalize_text(run_id),
-            thread_id=_normalize_text(thread_id),
-            lane_id=_normalize_lane_id(lane_id),
-            from_state=_normalize_text(from_state),
-            to_state=_normalize_text(to_state),
-            reason=_normalize_text(reason),
-            timestamp=_normalize_text(timestamp) or stamp,
-            idempotency_key=_normalize_text(idempotency_key),
-            subscription_id=_normalize_text(subscription_id),
-            timer_id=_normalize_text(timer_id),
-            event_id=_normalize_text(event_id),
-            event_type=_normalize_text(event_type),
-            event_data=dict(event_data or {}),
-            metadata=dict(metadata or {}),
-        )
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "PmaAutomationWakeup":
-        wakeup_id = _normalize_text(data.get("wakeup_id")) or str(uuid.uuid4())
-        created_at = _normalize_text(data.get("created_at")) or _iso_now()
-        updated_at = _normalize_text(data.get("updated_at")) or created_at
-        state = _normalize_text(data.get("state")) or "pending"
-        source = _normalize_text(data.get("source")) or "automation"
-        event_data_raw = data.get("event_data")
-        event_data: dict[str, Any] = (
-            dict(event_data_raw) if isinstance(event_data_raw, dict) else {}
-        )
-        metadata_raw = data.get("metadata")
-        metadata: dict[str, Any] = (
-            dict(metadata_raw) if isinstance(metadata_raw, dict) else {}
-        )
-        return cls(
-            wakeup_id=wakeup_id,
-            created_at=created_at,
-            updated_at=updated_at,
-            state=state.lower(),
-            dispatched_at=_normalize_text(data.get("dispatched_at")),
-            source=source,
-            repo_id=_normalize_text(data.get("repo_id")),
-            run_id=_normalize_text(data.get("run_id")),
-            thread_id=_normalize_text(data.get("thread_id")),
-            lane_id=_normalize_lane_id(data.get("lane_id")),
-            from_state=_normalize_text(data.get("from_state")),
-            to_state=_normalize_text(data.get("to_state")),
-            reason=_normalize_text(data.get("reason")),
-            timestamp=_normalize_text(data.get("timestamp")),
-            idempotency_key=_normalize_text(data.get("idempotency_key")),
-            subscription_id=_normalize_text(data.get("subscription_id")),
-            timer_id=_normalize_text(data.get("timer_id")),
-            event_id=_normalize_text(data.get("event_id")),
-            event_type=_normalize_text(data.get("event_type")),
-            event_data=event_data,
-            metadata=metadata,
-        )
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
-
-
 class PmaAutomationStore:
     def __init__(self, hub_root: Path, *, durable: bool = True) -> None:
         self._hub_root = hub_root
         self._durable = durable
-        self._path = (
-            hub_root / ".codex-autorunner" / "pma" / PMA_AUTOMATION_STORE_FILENAME
-        )
+        self._persistence = PmaAutomationPersistence(hub_root, durable=durable)
+        self._mirror = PmaAutomationMirror(hub_root, durable=durable)
 
     @property
     def path(self) -> Path:
-        return self._path
+        return self._persistence.path
 
     def _lock_path(self) -> Path:
-        return lock_path_for(self._path)
+        return self._persistence._lock_path()
 
     def load(self) -> dict[str, Any]:
-        with file_lock(self._lock_path()):
-            state = self._load_unlocked()
-            if state is not None:
-                return state
-            state = default_pma_automation_state()
-            return state
+        return self._persistence.load()
 
     def _load_unlocked(self) -> Optional[dict[str, Any]]:
-        try:
-            with open_orchestration_sqlite(
-                self._hub_root,
-                durable=self._durable,
-                migrate=False,
-            ) as conn:
-                subscriptions = self._load_subscriptions_from_sqlite(conn)
-                timers = self._load_timers_from_sqlite(conn)
-                wakeups = self._load_wakeups_from_sqlite(conn)
-        except sqlite3.OperationalError as exc:
-            if "no such table" not in str(exc).lower():
-                raise
-            return None
-        updated_values: list[str] = []
-        updated_values.extend(entry.updated_at for entry in subscriptions)
-        updated_values.extend(entry.updated_at for entry in timers)
-        updated_values.extend(entry.updated_at for entry in wakeups)
-        state = default_pma_automation_state()
-        if updated_values:
-            state["updated_at"] = max(updated_values)
-        state["subscriptions"] = [entry.to_dict() for entry in subscriptions]
-        state["timers"] = [entry.to_dict() for entry in timers]
-        state["wakeups"] = [entry.to_dict() for entry in wakeups]
-        return state
+        return self._persistence._load_unlocked()
 
     def _load_structured_unlocked(
         self,
@@ -518,15 +142,7 @@ class PmaAutomationStore:
         list[PmaAutomationTimer],
         list[PmaAutomationWakeup],
     ]:
-        state = self._load_unlocked()
-        if state is None:
-            state = default_pma_automation_state()
-        return (
-            state,
-            self._normalize_subscriptions(state.get("subscriptions")),
-            self._normalize_timers(state.get("timers")),
-            self._normalize_wakeups(state.get("wakeups")),
-        )
+        return self._persistence._load_structured_unlocked()
 
     def _save_structured_unlocked(
         self,
@@ -535,453 +151,9 @@ class PmaAutomationStore:
         timers: list[PmaAutomationTimer],
         wakeups: list[PmaAutomationWakeup],
     ) -> None:
-        subscription_ids = {
-            entry.subscription_id
-            for entry in subscriptions
-            if entry.subscription_id.strip()
-        }
-        filtered_timers = [
-            entry
-            for entry in timers
-            if entry.subscription_id is None
-            or entry.subscription_id in subscription_ids
-        ]
-        filtered_wakeups = [
-            entry
-            for entry in wakeups
-            if entry.subscription_id is None
-            or entry.subscription_id in subscription_ids
-        ]
-        dropped_timers = len(timers) - len(filtered_timers)
-        dropped_wakeups = len(wakeups) - len(filtered_wakeups)
-        if dropped_timers or dropped_wakeups:
-            logger.warning(
-                "Dropping orphaned automation rows before save (timers=%s, wakeups=%s)",
-                dropped_timers,
-                dropped_wakeups,
-            )
-        state["updated_at"] = _iso_now()
-        state["subscriptions"] = [entry.to_dict() for entry in subscriptions]
-        state["timers"] = [entry.to_dict() for entry in filtered_timers]
-        state["wakeups"] = [entry.to_dict() for entry in filtered_wakeups]
-        with open_orchestration_sqlite(self._hub_root, durable=self._durable) as conn:
-            with conn:
-                conn.execute("DELETE FROM orch_automation_wakeups")
-                conn.execute("DELETE FROM orch_automation_timers")
-                conn.execute("DELETE FROM orch_automation_subscriptions")
-                for subscription in subscriptions:
-                    conn.execute(
-                        """
-                        INSERT INTO orch_automation_subscriptions (
-                            subscription_id,
-                            event_types_json,
-                            repo_id,
-                            run_id,
-                            thread_target_id,
-                            binding_id,
-                            lane_id,
-                            from_state,
-                            to_state,
-                            notify_once,
-                            state,
-                            match_count,
-                            metadata_json,
-                            created_at,
-                            updated_at,
-                            disabled_at,
-                            reason_text,
-                            idempotency_key,
-                            max_matches
-                        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                        """,
-                        (
-                            subscription.subscription_id,
-                            json.dumps(subscription.event_types),
-                            subscription.repo_id,
-                            subscription.run_id,
-                            subscription.thread_id,
-                            None,
-                            subscription.lane_id,
-                            subscription.from_state,
-                            subscription.to_state,
-                            1 if subscription.max_matches == 1 else 0,
-                            subscription.state,
-                            subscription.match_count,
-                            json.dumps(subscription.metadata),
-                            subscription.created_at,
-                            subscription.updated_at,
-                            (
-                                subscription.updated_at
-                                if subscription.state != "active"
-                                else None
-                            ),
-                            subscription.reason,
-                            subscription.idempotency_key,
-                            subscription.max_matches,
-                        ),
-                    )
-                for timer in filtered_timers:
-                    conn.execute(
-                        """
-                        INSERT INTO orch_automation_timers (
-                            timer_id,
-                            subscription_id,
-                            repo_id,
-                            run_id,
-                            thread_target_id,
-                            timer_kind,
-                            schedule_key,
-                            available_at,
-                            payload_json,
-                            state,
-                            created_at,
-                            updated_at,
-                            fired_at,
-                            reason_text,
-                            idempotency_key,
-                            idle_seconds
-                        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                        """,
-                        (
-                            timer.timer_id,
-                            timer.subscription_id,
-                            timer.repo_id,
-                            timer.run_id,
-                            timer.thread_id,
-                            timer.timer_type,
-                            timer.subscription_id or timer.idempotency_key,
-                            timer.due_at,
-                            json.dumps(
-                                {
-                                    "metadata": timer.metadata,
-                                    "from_state": timer.from_state,
-                                    "to_state": timer.to_state,
-                                }
-                            ),
-                            timer.state,
-                            timer.created_at,
-                            timer.updated_at,
-                            timer.fired_at,
-                            timer.reason,
-                            timer.idempotency_key,
-                            timer.idle_seconds,
-                        ),
-                    )
-                for wakeup in filtered_wakeups:
-                    conn.execute(
-                        """
-                        INSERT INTO orch_automation_wakeups (
-                            wakeup_id,
-                            subscription_id,
-                            repo_id,
-                            run_id,
-                            thread_target_id,
-                            lane_id,
-                            wakeup_kind,
-                            state,
-                            available_at,
-                            claimed_at,
-                            completed_at,
-                            reason_text,
-                            payload_json,
-                            created_at,
-                            updated_at,
-                            dispatched_at,
-                            timestamp,
-                            idempotency_key,
-                            timer_id,
-                            event_id,
-                            event_type
-                        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                        """,
-                        (
-                            wakeup.wakeup_id,
-                            wakeup.subscription_id,
-                            wakeup.repo_id,
-                            wakeup.run_id,
-                            wakeup.thread_id,
-                            wakeup.lane_id,
-                            wakeup.source,
-                            wakeup.state,
-                            wakeup.timestamp,
-                            None,
-                            None,
-                            wakeup.reason,
-                            json.dumps(
-                                {
-                                    "metadata": wakeup.metadata,
-                                    "event_data": wakeup.event_data,
-                                    "from_state": wakeup.from_state,
-                                    "to_state": wakeup.to_state,
-                                }
-                            ),
-                            wakeup.created_at,
-                            wakeup.updated_at,
-                            wakeup.dispatched_at,
-                            wakeup.timestamp,
-                            wakeup.idempotency_key,
-                            wakeup.timer_id,
-                            wakeup.event_id,
-                            wakeup.event_type,
-                        ),
-                    )
-
-    @staticmethod
-    def _json_load(raw: str | None) -> dict[str, Any]:
-        if not isinstance(raw, str) or not raw.strip():
-            return {}
-        try:
-            parsed = json.loads(raw)
-        except json.JSONDecodeError:
-            return {}
-        return parsed if isinstance(parsed, dict) else {}
-
-    @staticmethod
-    def _json_load_list(raw: str | None) -> list[str]:
-        if not isinstance(raw, str) or not raw.strip():
-            return []
-        try:
-            parsed = json.loads(raw)
-        except json.JSONDecodeError:
-            return []
-        return [str(item).lower() for item in parsed if isinstance(item, str)]
-
-    def _row_to_subscription(self, row: Any) -> PmaLifecycleSubscription:
-        return PmaLifecycleSubscription(
-            subscription_id=str(row["subscription_id"]),
-            created_at=str(row["created_at"]),
-            updated_at=str(row["updated_at"]),
-            state=str(row["state"] or "active"),
-            event_types=self._json_load_list(row["event_types_json"]),
-            repo_id=row["repo_id"],
-            run_id=row["run_id"],
-            thread_id=row["thread_target_id"],
-            lane_id=str(row["lane_id"] or DEFAULT_PMA_LANE_ID),
-            from_state=row["from_state"],
-            to_state=row["to_state"],
-            reason=row["reason_text"],
-            idempotency_key=row["idempotency_key"],
-            max_matches=(
-                int(row["max_matches"]) if row["max_matches"] is not None else None
-            ),
-            match_count=int(row["match_count"] or 0),
-            metadata=self._json_load(row["metadata_json"]),
+        self._persistence._save_structured_unlocked(
+            state, subscriptions, timers, wakeups
         )
-
-    def _row_to_timer(self, row: Any) -> PmaAutomationTimer:
-        payload = self._json_load(row["payload_json"])
-        metadata = cast(
-            dict[str, Any],
-            (
-                payload.get("metadata")
-                if isinstance(payload.get("metadata"), dict)
-                else {}
-            ),
-        )
-        return PmaAutomationTimer(
-            timer_id=str(row["timer_id"]),
-            due_at=str(row["available_at"]),
-            created_at=str(row["created_at"]),
-            updated_at=str(row["updated_at"]),
-            state=str(row["state"] or "pending"),
-            fired_at=row["fired_at"],
-            timer_type=str(row["timer_kind"] or TIMER_TYPE_ONE_SHOT),
-            idle_seconds=(
-                int(row["idle_seconds"]) if row["idle_seconds"] is not None else None
-            ),
-            subscription_id=row["subscription_id"],
-            repo_id=row["repo_id"],
-            run_id=row["run_id"],
-            thread_id=row["thread_target_id"],
-            lane_id=_normalize_lane_id(payload.get("lane_id")),
-            from_state=_normalize_text(payload.get("from_state")),
-            to_state=_normalize_text(payload.get("to_state")),
-            reason=row["reason_text"],
-            idempotency_key=row["idempotency_key"],
-            metadata=metadata,
-        )
-
-    def _row_to_wakeup(self, row: Any) -> PmaAutomationWakeup:
-        payload = self._json_load(row["payload_json"])
-        event_data = cast(
-            dict[str, Any],
-            (
-                payload.get("event_data")
-                if isinstance(payload.get("event_data"), dict)
-                else {}
-            ),
-        )
-        metadata = cast(
-            dict[str, Any],
-            (
-                payload.get("metadata")
-                if isinstance(payload.get("metadata"), dict)
-                else {}
-            ),
-        )
-        return PmaAutomationWakeup(
-            wakeup_id=str(row["wakeup_id"]),
-            created_at=str(row["created_at"]),
-            updated_at=str(row["updated_at"]),
-            state=str(row["state"] or "pending"),
-            dispatched_at=row["dispatched_at"],
-            source=str(row["wakeup_kind"] or "automation"),
-            repo_id=row["repo_id"],
-            run_id=row["run_id"],
-            thread_id=row["thread_target_id"],
-            lane_id=str(row["lane_id"] or DEFAULT_PMA_LANE_ID),
-            from_state=_normalize_text(payload.get("from_state")),
-            to_state=_normalize_text(payload.get("to_state")),
-            reason=row["reason_text"],
-            timestamp=row["timestamp"],
-            idempotency_key=row["idempotency_key"],
-            subscription_id=row["subscription_id"],
-            timer_id=row["timer_id"],
-            event_id=row["event_id"],
-            event_type=row["event_type"],
-            event_data=event_data,
-            metadata=metadata,
-        )
-
-    def _load_subscriptions_from_sqlite(
-        self,
-        conn: Any,
-    ) -> list[PmaLifecycleSubscription]:
-        rows = conn.execute(
-            """
-            SELECT *
-              FROM orch_automation_subscriptions
-             ORDER BY created_at ASC, subscription_id ASC
-            """
-        ).fetchall()
-        return [self._row_to_subscription(row) for row in rows]
-
-    def _load_timers_from_sqlite(self, conn: Any) -> list[PmaAutomationTimer]:
-        rows = conn.execute(
-            """
-            SELECT *
-              FROM orch_automation_timers
-             ORDER BY created_at ASC, timer_id ASC
-            """
-        ).fetchall()
-        return [self._row_to_timer(row) for row in rows]
-
-    def _load_wakeups_from_sqlite(self, conn: Any) -> list[PmaAutomationWakeup]:
-        rows = conn.execute(
-            """
-            SELECT *
-              FROM orch_automation_wakeups
-             ORDER BY created_at ASC, wakeup_id ASC
-            """
-        ).fetchall()
-        return [self._row_to_wakeup(row) for row in rows]
-
-    def _insert_subscription_row(
-        self, conn: Any, subscription: PmaLifecycleSubscription
-    ) -> None:
-        conn.execute(
-            """
-            INSERT INTO orch_automation_subscriptions (
-                subscription_id, event_types_json, repo_id, run_id,
-                thread_target_id, binding_id, lane_id, from_state, to_state,
-                notify_once, state, match_count, metadata_json,
-                created_at, updated_at, disabled_at,
-                reason_text, idempotency_key, max_matches
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-            (
-                subscription.subscription_id,
-                json.dumps(subscription.event_types),
-                subscription.repo_id,
-                subscription.run_id,
-                subscription.thread_id,
-                None,
-                subscription.lane_id,
-                subscription.from_state,
-                subscription.to_state,
-                1 if subscription.max_matches == 1 else 0,
-                subscription.state,
-                subscription.match_count,
-                json.dumps(subscription.metadata),
-                subscription.created_at,
-                subscription.updated_at,
-                (subscription.updated_at if subscription.state != "active" else None),
-                subscription.reason,
-                subscription.idempotency_key,
-                subscription.max_matches,
-            ),
-        )
-
-    def _insert_timer_row(self, conn: Any, timer: PmaAutomationTimer) -> None:
-        conn.execute(
-            """
-            INSERT INTO orch_automation_timers (
-                timer_id, subscription_id, repo_id, run_id,
-                thread_target_id, timer_kind, schedule_key, available_at,
-                payload_json, state, created_at, updated_at,
-                fired_at, reason_text, idempotency_key, idle_seconds
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-            (
-                timer.timer_id,
-                timer.subscription_id,
-                timer.repo_id,
-                timer.run_id,
-                timer.thread_id,
-                timer.timer_type,
-                timer.subscription_id or timer.idempotency_key,
-                timer.due_at,
-                json.dumps(
-                    {
-                        "metadata": timer.metadata,
-                        "from_state": timer.from_state,
-                        "to_state": timer.to_state,
-                    }
-                ),
-                timer.state,
-                timer.created_at,
-                timer.updated_at,
-                timer.fired_at,
-                timer.reason,
-                timer.idempotency_key,
-                timer.idle_seconds,
-            ),
-        )
-
-    def _find_active_subscription_by_key(
-        self, conn: Any, key: str
-    ) -> Optional[PmaLifecycleSubscription]:
-        rows = conn.execute(
-            """
-            SELECT *
-              FROM orch_automation_subscriptions
-             WHERE idempotency_key = ? AND state = 'active'
-             LIMIT 1
-            """,
-            (key,),
-        ).fetchall()
-        return self._row_to_subscription(rows[0]) if rows else None
-
-    def _find_pending_timer_by_key(
-        self, conn: Any, key: str
-    ) -> Optional[PmaAutomationTimer]:
-        rows = conn.execute(
-            """
-            SELECT *
-              FROM orch_automation_timers
-             WHERE idempotency_key = ? AND state = 'pending'
-             LIMIT 1
-            """,
-            (key,),
-        ).fetchall()
-        return self._row_to_timer(rows[0]) if rows else None
-
-    def _subscription_id_exists(self, conn: Any, sub_id: str) -> bool:
-        row = conn.execute(
-            "SELECT 1 FROM orch_automation_subscriptions WHERE subscription_id = ?",
-            (sub_id,),
-        ).fetchone()
-        return row is not None
 
     def _normalize_subscriptions(self, value: Any) -> list[PmaLifecycleSubscription]:
         out: list[PmaLifecycleSubscription] = []
@@ -1102,98 +274,21 @@ class PmaAutomationStore:
         )
         wakeup.metadata["dispatch_decision"] = pma_dispatch_decision_to_dict(decision)
 
-    @staticmethod
-    def _lifecycle_sub_to_domain(
-        entry: PmaLifecycleSubscription,
-    ) -> PmaSubscription:
-        return PmaSubscription(
-            subscription_id=entry.subscription_id,
-            created_at=entry.created_at,
-            updated_at=entry.updated_at,
-            state=entry.state,
-            event_types=tuple(entry.event_types),
-            repo_id=entry.repo_id,
-            run_id=entry.run_id,
-            thread_id=entry.thread_id,
-            lane_id=entry.lane_id,
-            from_state=entry.from_state,
-            to_state=entry.to_state,
-            reason=entry.reason,
-            idempotency_key=entry.idempotency_key,
-            max_matches=entry.max_matches,
-            match_count=entry.match_count,
-            metadata=dict(entry.metadata),
-        )
-
-    @staticmethod
-    def _store_timer_to_domain(entry: PmaAutomationTimer) -> PmaTimer:
-        return PmaTimer(
-            timer_id=entry.timer_id,
-            due_at=entry.due_at,
-            created_at=entry.created_at,
-            updated_at=entry.updated_at,
-            state=entry.state,
-            fired_at=entry.fired_at,
-            timer_type=entry.timer_type,
-            idle_seconds=entry.idle_seconds,
-            subscription_id=entry.subscription_id,
-            repo_id=entry.repo_id,
-            run_id=entry.run_id,
-            thread_id=entry.thread_id,
-            lane_id=entry.lane_id,
-            from_state=entry.from_state,
-            to_state=entry.to_state,
-            reason=entry.reason,
-            idempotency_key=entry.idempotency_key,
-            metadata=dict(entry.metadata or {}),
-        )
-
-    @staticmethod
-    def _store_wakeup_to_domain(entry: PmaAutomationWakeup) -> PmaWakeup:
-        return PmaWakeup(
-            wakeup_id=entry.wakeup_id,
-            created_at=entry.created_at,
-            updated_at=entry.updated_at,
-            state=entry.state,
-            dispatched_at=entry.dispatched_at,
-            source=entry.source,
-            repo_id=entry.repo_id,
-            run_id=entry.run_id,
-            thread_id=entry.thread_id,
-            lane_id=entry.lane_id,
-            from_state=entry.from_state,
-            to_state=entry.to_state,
-            reason=entry.reason,
-            timestamp=entry.timestamp,
-            idempotency_key=entry.idempotency_key,
-            subscription_id=entry.subscription_id,
-            timer_id=entry.timer_id,
-            event_id=entry.event_id,
-            event_type=entry.event_type,
-            event_data=dict(entry.event_data or {}),
-            metadata=dict(entry.metadata or {}),
-        )
-
-    @staticmethod
-    def _apply_domain_timer_to_store(
-        store_timer: PmaAutomationTimer, domain_timer: PmaTimer
-    ) -> None:
-        store_timer.due_at = domain_timer.due_at
-        store_timer.updated_at = domain_timer.updated_at
-        store_timer.state = domain_timer.state
-        store_timer.fired_at = domain_timer.fired_at
-        store_timer.idle_seconds = domain_timer.idle_seconds
-        store_timer.reason = domain_timer.reason
-        store_timer.metadata = dict(domain_timer.metadata or {})
-
-    @staticmethod
-    def _apply_domain_wakeup_to_store(
-        store_wakeup: PmaAutomationWakeup, domain_wakeup: PmaWakeup
-    ) -> None:
-        store_wakeup.state = domain_wakeup.state
-        store_wakeup.dispatched_at = domain_wakeup.dispatched_at
-        store_wakeup.updated_at = domain_wakeup.updated_at
-        store_wakeup.metadata = dict(domain_wakeup.metadata or {})
+    _lifecycle_sub_to_domain = staticmethod(
+        PmaAutomationDomainTranslator._lifecycle_sub_to_domain
+    )
+    _store_timer_to_domain = staticmethod(
+        PmaAutomationDomainTranslator._store_timer_to_domain
+    )
+    _store_wakeup_to_domain = staticmethod(
+        PmaAutomationDomainTranslator._store_wakeup_to_domain
+    )
+    _apply_domain_timer_to_store = staticmethod(
+        PmaAutomationDomainTranslator._apply_domain_timer_to_store
+    )
+    _apply_domain_wakeup_to_store = staticmethod(
+        PmaAutomationDomainTranslator._apply_domain_wakeup_to_store
+    )
 
     def _apply_reduce_result(
         self,
@@ -1204,42 +299,15 @@ class PmaAutomationStore:
         *,
         compute_dispatch: bool = True,
     ) -> list[PmaAutomationWakeup]:
-        sub_by_id = {entry.subscription_id: entry for entry in subscriptions}
-        now = _iso_now()
-        for domain_sub in result.subscriptions:
-            existing = sub_by_id.get(domain_sub.subscription_id)
-            if existing is None:
-                continue
-            if existing.match_count != domain_sub.match_count:
-                existing.match_count = domain_sub.match_count
-                existing.updated_at = now
-            if existing.state != domain_sub.state:
-                existing.state = domain_sub.state
-                existing.updated_at = now
-        appended: list[PmaAutomationWakeup] = []
-        for intent in result.wakeup_intents:
-            wakeup = PmaAutomationWakeup.create(
-                source=intent.source,
-                repo_id=intent.repo_id,
-                run_id=intent.run_id,
-                thread_id=intent.thread_id,
-                lane_id=intent.lane_id,
-                from_state=intent.from_state,
-                to_state=intent.to_state,
-                reason=intent.reason,
-                timestamp=timestamp,
-                idempotency_key=intent.idempotency_key,
-                subscription_id=intent.subscription_id,
-                event_type=intent.event_type,
-                event_id=intent.event_id,
-                event_data=intent.event_data,
-                metadata=intent.metadata,
-            )
-            if compute_dispatch:
-                self._compute_dispatch_decision_for_wakeup(wakeup)
-            wakeups.append(wakeup)
-            appended.append(wakeup)
-        return appended
+        return PmaAutomationDomainTranslator._apply_reduce_result(
+            subscriptions,
+            wakeups,
+            result,
+            timestamp,
+            compute_dispatch=(
+                self._compute_dispatch_decision_for_wakeup if compute_dispatch else None
+            ),
+        )
 
     @staticmethod
     def _normalize_subscription_event_types(
@@ -1536,7 +604,9 @@ class PmaAutomationStore:
             ) as conn:
                 with conn:
                     if key is not None:
-                        existing = self._find_active_subscription_by_key(conn, key)
+                        existing = self._persistence._find_active_subscription_by_key(
+                            conn, key
+                        )
                         if existing is not None:
                             created = existing
                             deduped = True
@@ -1558,7 +628,7 @@ class PmaAutomationStore:
                             ),
                             metadata=resolved_metadata,
                         )
-                        self._insert_subscription_row(conn, created)
+                        self._persistence._insert_subscription_row(conn, created)
         self._mirror_subscription_rule(created)
         return created, deduped
 
@@ -1692,16 +762,7 @@ class PmaAutomationStore:
             return changed
 
     def _mirror_subscription_rule(self, subscription: PmaLifecycleSubscription) -> None:
-        try:
-            mirror_pma_subscription_rule(
-                AutomationStore(self._hub_root, durable=self._durable),
-                subscription=subscription,
-            )
-        except (sqlite3.Error, OSError, ValueError, TypeError, RuntimeError):
-            logger.exception(
-                "Failed to mirror PMA subscription into unified automation rule: %s",
-                getattr(subscription, "subscription_id", None),
-            )
+        self._mirror.mirror_subscription_rule(subscription)
 
     def purge_subscription(
         self, subscription_id: str, *, require_inactive: bool = True
@@ -1767,7 +828,7 @@ class PmaAutomationStore:
                     query += " WHERE state = ?"
                     params = (target_state,)
                 rows = conn.execute(query, params).fetchall()
-                removed = [self._row_to_subscription(row) for row in rows]
+                removed = [self._persistence._row_to_subscription(row) for row in rows]
                 if removed and not dry_run:
                     removed_ids = [entry.subscription_id for entry in removed]
                     total_orphaned_timers = 0
@@ -1911,14 +972,16 @@ class PmaAutomationStore:
                 with conn:
                     normalized_subscription_id = _normalize_text(subscription_id)
                     if normalized_subscription_id is not None:
-                        if not self._subscription_id_exists(
+                        if not self._persistence._subscription_id_exists(
                             conn, normalized_subscription_id
                         ):
                             raise ValueError(
                                 f"Unknown subscription_id: {normalized_subscription_id}"
                             )
                     if key is not None:
-                        existing = self._find_pending_timer_by_key(conn, key)
+                        existing = self._persistence._find_pending_timer_by_key(
+                            conn, key
+                        )
                         if existing is not None:
                             created = existing
                             deduped = True
@@ -1938,7 +1001,7 @@ class PmaAutomationStore:
                             idempotency_key=key,
                             metadata=metadata,
                         )
-                        self._insert_timer_row(conn, created)
+                        self._persistence._insert_timer_row(conn, created)
         self._mirror_timer_schedule(created)
         return created, deduped
 
@@ -2155,16 +1218,7 @@ class PmaAutomationStore:
         return {"status": "ok", "timer_id": target_id, "touched": False}
 
     def _mirror_timer_schedule(self, timer: PmaAutomationTimer) -> None:
-        try:
-            mirror_pma_timer_schedule(
-                AutomationStore(self._hub_root, durable=self._durable),
-                timer=timer,
-            )
-        except (sqlite3.Error, OSError, ValueError, TypeError, RuntimeError):
-            logger.exception(
-                "Failed to mirror PMA timer into unified automation schedule: %s",
-                getattr(timer, "timer_id", None),
-            )
+        self._mirror.mirror_timer_schedule(timer)
 
     def dequeue_due_timers(
         self,
